requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-environ==0.9.0
django-filter==22.1
django-countries==7.4.2
Pillow==9.3.0
psycopg2-binary==2.9.5
flake8==6.0.0
black==22.10.0
isort==5.10.1
django-phonenumber-field==7.0.0
phonenumbers==8.13.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
Faker==15.3.4
pytest-cov==2.12.1
drf-yasg==1.21.4
django-rq==2.6.0
djangorestframework-simplejwt==5.0.0
google-api-python-client==2.70.0
facebook-sdk==3.1.0
python-twitter==3.5
django-taggit==3.1.0